TST: collect libhashtable tests
